Add method to get directions in random order.
I tried using it on level 1, resulting in same maxDepth value and 10 ms
slower in average (on 10 runs). However, this slowdown might be due to the
shuffling operations.

Even though not activated in this version, it might be a good idea
thinking about how we select next direction to walk.