Add benchmarking stub and refactor Solver.
The reason that I refactored Solver to let solve() take a board is that
it makes the benchmarking somewhat smoother. However, we might want to
have better code in the solvers than in the benchmarking. Therefore I do
not refactor the Puller solver.

For that reason, the current tree does not compile. However, the
benchmarking is working for the IDS solver.